Improve the performances of the consumption migration task by using the charging stations instead updating each consumption
