+++ src/migration/MigrationHandler.ts
@@ -19,7 +19,7 @@ import UpdateChargingStationTemplatesTask from './tasks/UpdateChargingStationTem
 import CleanupOrphanBadgeTask from './tasks/CleanupOrphanBadgeTask';
 import AddIssuerFieldTask from './tasks/AddIssuerFieldTask';
 import AddLastChangePropertiesToBadgeTask from './tasks/AddLastChangePropertiesToBadgeTask';
-import AddLimitToConsumptions from './tasks/AddLimitToConsumptions';
+import AddLimitToConsumptionsTask from './tasks/AddLimitToConsumptionsTask';
 import AddActivePropertyToTagsTask from './tasks/AddActivePropertyToTagsTask';
 
 export default class MigrationHandler {
@@ -55,7 +55,7 @@ export default class MigrationHandler {
       currentMigrationTasks.push(new AddIssuerFieldTask());
       currentMigrationTasks.push(new CleanupOrphanBadgeTask());
       currentMigrationTasks.push(new AddLastChangePropertiesToBadgeTask());
-      currentMigrationTasks.push(new AddLimitToConsumptions());
+      currentMigrationTasks.push(new AddLimitToConsumptionsTask());
       currentMigrationTasks.push(new AddActivePropertyToTagsTask());
 
       // Get the already done migrations from the DB
